Add option to use Nickname instead of Username
Currently matterircd uses Username as the nick handler for every
user, regardless if such user have a Nickname set. There is an inherent
restriction on using the Nickname - they may contain non-ascii chars,
breaking compatibility with IRC clients.

This patch adds the config option PreferNickname - if set, we check
Nickname and if it's not empty and doesn't contain any non-ascii
character, matterircd uses the Nickname instead of Username.
It's a Mattermost-only option for the time being.

Signed-off-by: Guilherme G. Piccoli <gpiccoli@canonical.com>